Add clj-toolbox.prelude/not-empty?
This was prompted by clj-kondo constantly complaining about
(not (empty? x)) being an anti-pattern since empty? is apparently just
(not (seq x)) on the inside and they wanted to avoid the double not.

But their suggestion of using seq for this purpose is not very readable
by itself. not-empty? is more clear to me.